requirements.txt
joblib
keras
keras_nightly
matplotlib
numpy
pandas
Pillow
psutil
scikit_learn
seaborn
tensorflow
tf_nightly
tqdm
Flask
Flask-cors